Use a memoized key lookup for rules
On my project, this makes ESLint as a whole roughly 8% faster, and it
seems to result in consistently faster benchmarks.

Note a behavior change - member expressions used to find the first
matching rule, while they now favor the more general rule if relevant.
(I believe this change is an improvement: "Promise is not supported in
IE 10" is a more helpful message than "Promise.reject() is not supported
in IE 10".)